[trusty][arch][x86-64] Enable 64bit TA and Lib for trusty
1. re-implemented the 64bit syscall
2. keep align the type define of the ipc structures
3. replaced the ELF32*_ to ELF64* to parse the 64bit elf files.

+++ lib/trusty/include/lib/trusty/ipc_msg.h
@@ -64,7 +64,7 @@ typedef struct ipc_msg_user {
 } ipc_msg_user_t;
 
 typedef struct ipc_msg_info {
-	uint32_t	len;
+	size_t	len;
 	uint32_t	id;
 } ipc_msg_info_t;
